LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-named-scroll-progress-timeline.tentative-expected.txt


PASS scroll-timeline-name is referenceable in animation-timeline on the declaring element itself
PASS scroll-timeline-name is referenceable in animation-timeline on that element's descendants
FAIL scroll-timeline-name is not referenceable in animation-timeline on that element's siblings assert_equals: Animation with unknown timeline name holds current time at zero expected "50px" but got "none"
PASS scroll-timeline-name on an element which is not a scroll-container
FAIL Change in scroll-timeline-name to match animation timeline updates animation. assert_equals: expected null but got object "[object DocumentTimeline]"
FAIL Change in scroll-timeline-name to no longer match animation timeline updates animation. assert_equals: Failed to remove timeline expected null but got object "[object ScrollTimeline]"
FAIL Timeline lookup updates candidate when closer match available. assert_equals: Timeline not updated expected "B" but got "A"
PASS Timeline lookup updates candidate when match becomes available.
PASS scroll-timeline-axis is block
PASS scroll-timeline-axis is inline
PASS scroll-timeline-axis is x
PASS scroll-timeline-axis is y
FAIL scroll-timeline-axis is mutated assert_equals: expected "75" but got "25"
